web/requirements.txt
# Pyhon Environment
python-dotenv==0.19.2
python-dateutil==2.8.1
fonttools==4.29.1

# Django
Django==4.1.8
django-crispy-forms==1.14.0
django-floppyforms==1.9.0
django-formtools==2.3
django-guardian==2.4.0
django-livereload-server==0.4
django-map-widgets==0.4.0
djangorestframework==3.13.1
django-extensions
drf-extra-fields==3.4.0
drf-chunked-upload==0.5.1
drf-yasg==1.20.0
channels==4.0.0
daphne==4.0.0
requests==2.24.0
json2html==1.3.0
user-agents==2.2.0
pytz==2021.1
pandas==1.1.5
opencv-python==4.5.1.48
# numpy==1.20.0
numpy==1.26.0

# Drone
pymavlink==2.4.29

# Database
mysqlclient==2.1.0

# Lidar Point Cloud
open3d==0.16.0
itsdangerous==2.0.1

# PDF
reportlab==3.6.12
pdfkit==1.0.0
pdfrw==0.4 

# Monitoring system
psutil==5.8.0
GPUtil==1.4.0 

# Georeference
geopy==1.23.0

# Fire pretiction
geojson==2.5.0
scipy==1.5.4

# Weather 
pyserial==3.5

# Testing system
factory-boy==3.2.1
Faker==13.0.0

# Save Drones
sympy==1.9

# Mesh monitoring
pyodm==1.5.9


gunicorn

# for operation coverage polygons
alpha-shapes
shapely